+++ src/api/webcast.ts
@@ -1,6 +1,6 @@
 import { Rev } from '..';
 import type { RevClient } from '../rev-client';
-import { Webcast, GuestRegistration } from '../types/webcast';
+import { Webcast, GuestRegistration, Banner } from '../types/webcast';
 import { SearchRequest } from '../utils/request-utils';
 import { titleCase } from '../utils';
 import { PostEventReportRequest, RealtimeReportRequest } from './webcast-report-request';
@@ -14,8 +14,8 @@ type RealtimeSession<T extends Webcast.RealtimeRequest | undefined> = T extends 
 
 export default function webcastAPIFactory(rev: RevClient) {
     const webcastAPI = {
-        async list(options: Webcast.ListRequest = { }): Promise<Webcast[]> {
-            return rev.get('/api/v2/scheduled-events', options, { responseType: 'json' });
+        async list(options: Webcast.ListRequest = { }, requestOptions?: Rev.RequestOptions): Promise<Webcast[]> {
+            return rev.get('/api/v2/scheduled-events', options, { ...requestOptions, responseType: 'json' });
         },
         search(query: Webcast.SearchRequest, options?: Rev.SearchOptions<Webcast>): Rev.ISearchRequest<Webcast> {
             const searchDefinition: Rev.SearchDefinition<Webcast> = {
@@ -31,8 +31,8 @@ export default function webcastAPIFactory(rev: RevClient) {
             const { eventId } = await rev.post(`/api/v2/scheduled-events`, event);
             return eventId;
         },
-        async details(eventId: string): Promise<Webcast.Details> {
-            return rev.get(`/api/v2/scheduled-events/${eventId}`);
+        async details(eventId: string, requestOptions?: Rev.RequestOptions): Promise<Webcast.Details> {
+            return rev.get(`/api/v2/scheduled-events/${eventId}`, undefined, requestOptions);
         },
         async edit(eventId: string, event: Webcast.CreateRequest): Promise<void> {
             return rev.put(`/api/v2/scheduled-events/${eventId}`, event);
@@ -74,11 +74,11 @@ export default function webcastAPIFactory(rev: RevClient) {
             const query = (runNumber ?? -1) >= 0 ? { runNumber } : {};
             return rev.get(`/api/v2/scheduled-events/${eventId}/comments`, query, { responseType: 'json' });
         },
-        async status(eventId: string): Promise<Webcast.Status> {
-            return rev.get(`/api/v2/scheduled-events/${eventId}/status`);
+        async status(eventId: string, requestOptions?: Rev.RequestOptions): Promise<Webcast.Status> {
+            return rev.get(`/api/v2/scheduled-events/${eventId}/status`, undefined, requestOptions);
         },
-        async isPublic(eventId: string): Promise<boolean> {
-            const response = await rev.request('GET', `/api/v2/scheduled-events/${eventId}/is-public`, undefined, { throwHttpErrors: false, responseType: 'json' });
+        async isPublic(eventId: string, requestOptions?: Rev.RequestOptions): Promise<boolean> {
+            const response = await rev.request('GET', `/api/v2/scheduled-events/${eventId}/is-public`, undefined, { ...requestOptions, throwHttpErrors: false, responseType: 'json' });
             return response.statusCode !== 401 && response.body?.isPublic;
         },
         async playbackUrls(eventId: string, {ip, userAgent}: Webcast.PlaybackUrlRequest = { }, options?: Rev.RequestOptions): Promise<Webcast.PlaybackUrlsResponse> {
@@ -178,7 +178,24 @@ export default function webcastAPIFactory(rev: RevClient) {
         deleteGuestRegistration(eventId: string, registrationId: string): Promise<void> {
             return rev.delete(`/api/v2/scheduled-events/${eventId}/registrations/${registrationId}`);
         },
-
+        async listBanners(eventId: string): Promise<Banner[]> {
+            const {banners} = await rev.get(`/api/v2/scheduled-events/${eventId}/banners`);
+            return banners || [];
+        },
+        addBanner(eventId: string, banner: Banner.Request): Promise<Banner> {
+            return rev.post(`/api/v2/scheduled-events/${eventId}/banner`, banner);
+        },
+        setBannerStatus(eventId: string, bannerId: string, isEnabled: boolean): Promise<void> {
+            return rev.put(`/api/v2/scheduled-events/${eventId}/banner/${bannerId}/status`, {isEnabled});
+        },
+        updateBanner(eventId: string, banner: Banner): Promise<Banner> {
+            // separate id from the banner data
+            const {id, ...payload} = banner;
+            return rev.put(`/api/v2/scheduled-events/${eventId}/banner/${id}`, payload);
+        },
+        deleteBanner(eventId: string, bannerId: string): Promise<void> {
+            return rev.delete(`/api/v2/scheduled-events/${eventId}/banner/${bannerId}`);
+        }
     };
 
     return webcastAPI;