+++ src/index.ts
@@ -3,11 +3,13 @@ export * from './rev-error';
 export type * from './types';
 
 import {rateLimit} from './utils';
-import {getExtensionForMime, getMimeForExtension} from './utils/file-utils'
+import {getExtensionForMime, getMimeForExtension} from './utils/file-utils';
+import { setPolyfills } from './interop/polyfills';
 export const utils = {
     rateLimit,
     getExtensionForMime,
-    getMimeForExtension
+    getMimeForExtension,
+    setPolyfills
 };
 
 import { RevClient } from './rev-client';